requirements.txt
# GH Systems ABC - Consolidated Requirements
# Install with: pip install -r requirements.txt
# For optional extras, see requirements/ directory or install specific packages

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# Data Validation
pydantic>=2.0.0

# Graph Processing
networkx>=3.0

# HTTP Requests
requests>=2.31.0

# Caching
redis>=5.0.0

# Environment Variables
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0

# Cryptographic Signatures (REQUIRED for production)
cryptography>=41.0.0

# JWT Authentication (Security)
PyJWT>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: AI Ontology (install if needed)
# numpy>=1.24.0
# scikit-learn>=1.3.0
# torch>=2.0.0
# transformers>=4.30.0
# torch-geometric>=2.3.0
# dgl>=1.1.0
# openai>=1.0.0
# anthropic>=0.7.0
# pandas>=2.0.0

# Oracle Layer dependencies
python-bitcoinlib>=0.12.0  # Bitcoin RPC client (optional, for oracle)
websockets>=12.0  # WebSocket support for real-time streams

# Optional: Vector Database (install if needed)
# qdrant-client>=1.6.0
# faiss-cpu>=1.7.4
# sentence-transformers>=2.2.0

# Optional: Database (install if needed)
# psycopg2-binary>=2.9.0  # PostgreSQL adapter for dashboard
